Create Event and Registration entity classes
Event abstract class is a base class for CampEvent, TournamentEvent and
ExamEvent classes which represent types of events that may occur in
the club.

Because user has to be able to choose some custom options while registering
to an event, Registration classes have been created to represent a
user's registration for a particular event with custom options like
meeting participation, accomodation, room type, etc.