Send customSessionId by authenticateUser endpoint
A customSessionId value is generated by UUID.randomUUID() method in
authenticateUser() endpoint and sent with JwtResponse with the rest of
user's data.

Any kind of unique session id value is necessary for image storage handling.
When an administrator is setting an image while creating a new event,
the image is is being sent to the server and temporarily saved.
Because there might be several administrators working at the same time,
image sent must be saved in a directory named with a kind of unique value,
which may be a session id - generated by a server side.