Fix "add" and "update" endpoints in BranchChief, Club and Rank controller classes
An "update" method has been modified so that first it checks if a given entity exists
by name. If so - returns a ResponseEntity with an error code, otherwise saves
an entity.
As the code from "update" and "save" methods is now exactly the same, it was moved
into a private method saveIfNotExist.